+++ R/eval-rename.R
@@ -6,14 +6,20 @@ eval_rename <- function(expr,
                         ...,
                         strict = TRUE,
                         name_spec = NULL,
+                        allow_predicates = TRUE,
                         error_call = caller_env()) {
   check_dots_empty()
+
+  allow_predicates <- allow_predicates && tidyselect_data_has_predicates(data)
+  data <- tidyselect_data_proxy(data)
+
   rename_impl(
     data,
     names(data),
     as_quosure(expr, env),
     strict = strict,
     name_spec = name_spec,
+    allow_predicates = allow_predicates,
     error_call = error_call
   )
 }
@@ -24,6 +30,7 @@ rename_impl <- function(x,
                         sel,
                         strict = TRUE,
                         name_spec = NULL,
+                        allow_predicates = TRUE,
                         error_call) {
   if (is_null(names)) {
     cli::cli_abort("Can't rename an unnamed vector.", call = error_call)
@@ -36,6 +43,7 @@ rename_impl <- function(x,
     strict = strict,
     name_spec = name_spec,
     type = "rename",
+    allow_predicates = allow_predicates,
     error_call = error_call
   )